pyproject.toml
[project]
name = "srv-ai-outreach"
version = "2.0.0"
description = "SRV AI Email Outreach — CyberArc MSP (FastAPI + SQLModel)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.21",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.13.1",
]

[dependency-groups]
dev = [
    "pillow>=12.1.1",
]
